fix(sandbox): encode path segments and relocate ide deep-link helper
Address PR review:
- Encode each path segment with encodeURIComponent (keeping `/` and the
  drive-letter `:` literal) instead of encodeURI, which left `#`/`%`/`?`
  unescaped. A `#` in the path (e.g. a Windows home dir `C:\Users\John#Doe`)
  silently truncated the URL into a fragment, dropping `?windowId=_blank`.
- Add regression tests for `#` and literal `%` in the path.
- Move the helper from components/sandbox/preview/ to components/sandbox/
  since it is consumed from outside the preview tree (code-tab.tsx).
